+++ worker/src/RTC/Transport.cpp
@@ -27,6 +27,9 @@
 
 namespace RTC
 {
+	static size_t DefaultSctpSendBufferSize{ 262144 }; // 2^18.
+	static size_t MaxSctpSendBufferSize{ 268435456 };  // 2^28.
+
 	/* Instance methods. */
 
 	Transport::Transport(const std::string& id, Listener* listener, json& data)
@@ -90,6 +93,7 @@ namespace RTC
 
 			auto jsonNumSctpStreamsIt     = data.find("numSctpStreams");
 			auto jsonMaxSctpMessageSizeIt = data.find("maxSctpMessageSize");
+			auto jsonSctpSendBufferSizeIt = data.find("sctpSendBufferSize");
 			auto jsonIsDataChannelIt      = data.find("isDataChannel");
 
 			// numSctpStreams is mandatory.
@@ -135,6 +139,28 @@ namespace RTC
 
 			this->maxMessageSize = jsonMaxSctpMessageSizeIt->get<size_t>();
 
+			size_t sctpSendBufferSize;
+
+			// sctpSendBufferSize is optional.
+			if (jsonSctpSendBufferSizeIt != data.end())
+			{
+				if (!Utils::Json::IsPositiveInteger(*jsonSctpSendBufferSizeIt))
+				{
+					MS_THROW_TYPE_ERROR("wrong sctpSendBufferSize (not a number)");
+				}
+
+				sctpSendBufferSize = jsonSctpSendBufferSizeIt->get<size_t>();
+
+				if (sctpSendBufferSize > MaxSctpSendBufferSize)
+				{
+					MS_THROW_TYPE_ERROR("wrong sctpSendBufferSize (maximum value exceeded)");
+				}
+			}
+			else
+			{
+				sctpSendBufferSize = DefaultSctpSendBufferSize;
+			}
+
 			// isDataChannel is optional.
 			bool isDataChannel{ false };
 
@@ -142,8 +168,8 @@ namespace RTC
 				isDataChannel = jsonIsDataChannelIt->get<bool>();
 
 			// This may throw.
-			this->sctpAssociation =
-			  new RTC::SctpAssociation(this, os, mis, this->maxMessageSize, isDataChannel);
+			this->sctpAssociation = new RTC::SctpAssociation(
+			  this, os, mis, this->maxMessageSize, sctpSendBufferSize, isDataChannel);
 		}
 
 		// Create the RTCP timer.
@@ -1421,47 +1447,6 @@ namespace RTC
 
 		switch (notification->eventId)
 		{
-			case PayloadChannel::Notification::EventId::DATA_PRODUCER_SEND:
-			{
-				// This may throw.
-				RTC::DataProducer* dataProducer = GetDataProducerFromInternal(notification->internal);
-
-				if (dataProducer->GetType() != RTC::DataProducer::Type::DIRECT)
-				{
-					MS_THROW_ERROR("cannot send direct messages on this DataProducer");
-				}
-
-				auto jsonPpidIt = notification->data.find("ppid");
-
-				if (jsonPpidIt == notification->data.end() || !Utils::Json::IsPositiveInteger(*jsonPpidIt))
-				{
-					MS_THROW_TYPE_ERROR("invalid ppid");
-				}
-
-				auto ppid       = jsonPpidIt->get<uint32_t>();
-				auto len        = notification->payloadLen;
-				const auto* msg = notification->payload;
-
-				if (len > this->maxMessageSize)
-				{
-					MS_WARN_TAG(
-					  message,
-					  "given message exceeds maxMessageSize value [maxMessageSize:%zu, len:%zu]",
-					  len,
-					  this->maxMessageSize);
-
-					return;
-				}
-
-				// Pass the message to the DataProducer.
-				dataProducer->ReceiveMessage(ppid, msg, len);
-
-				// Increase receive transmission.
-				DataReceived(len);
-
-				break;
-			}
-
 			default:
 			{
 				MS_ERROR("unknown event '%s'", notification->event.c_str());
@@ -1868,7 +1853,11 @@ namespace RTC
 					{
 						auto* consumer = GetConsumerByMediaSsrc(feedback->GetMediaSsrc());
 
-						if (!consumer)
+						if (feedback->GetMediaSsrc() == RTC::RtpProbationSsrc)
+						{
+							break;
+						}
+						else if (!consumer)
 						{
 							MS_DEBUG_TAG(
 							  rtcp,
@@ -1903,7 +1892,11 @@ namespace RTC
 							auto& item     = *it;
 							auto* consumer = GetConsumerByMediaSsrc(item->GetSsrc());
 
-							if (!consumer)
+							if (item->GetSsrc() == RTC::RtpProbationSsrc)
+							{
+								continue;
+							}
+							else if (!consumer)
 							{
 								MS_DEBUG_TAG(
 								  rtcp,
@@ -1980,17 +1973,15 @@ namespace RTC
 				auto* consumer = GetConsumerByMediaSsrc(feedback->GetMediaSsrc());
 
 				// If no Consumer is found and this is not a Transport Feedback for the
-				// probation SSRC or any Consumer RTX SSRC ignore it.
+				// probation SSRC or any Consumer RTX SSRC, ignore it.
 				//
 				// clang-format off
 				if (
 					!consumer &&
+					feedback->GetMessageType() != RTC::RTCP::FeedbackRtp::MessageType::TCC &&
 					(
-						(feedback->GetMessageType() != RTC::RTCP::FeedbackRtp::MessageType::TCC) &&
-						(
-						 feedback->GetMediaSsrc() != RTC::RtpProbationSsrc ||
-						 !GetConsumerByRtxSsrc(feedback->GetMediaSsrc())
-						)
+						feedback->GetMediaSsrc() != RTC::RtpProbationSsrc ||
+						!GetConsumerByRtxSsrc(feedback->GetMediaSsrc())
 					)
 				)
 				// clang-format on
@@ -2009,6 +2000,18 @@ namespace RTC
 				{
 					case RTC::RTCP::FeedbackRtp::MessageType::NACK:
 					{
+						if (!consumer)
+						{
+							MS_DEBUG_TAG(
+							  rtcp,
+							  "no Consumer found for received NACK Feedback packet "
+							  "[sender ssrc:%" PRIu32 ", media ssrc:%" PRIu32 "]",
+							  feedback->GetMediaSsrc(),
+							  feedback->GetMediaSsrc());
+
+							break;
+						}
+
 						auto* nackPacket = static_cast<RTC::RTCP::FeedbackRtpNackPacket*>(packet);
 
 						consumer->ReceiveNack(nackPacket);
@@ -2419,7 +2422,7 @@ namespace RTC
 		  this, producer, mappedSsrc, worstRemoteFractionLost);
 	}
 
-	inline void Transport::OnConsumerSendRtpPacket(RTC::Consumer* /*consumer*/, RTC::RtpPacket* packet)
+	inline void Transport::OnConsumerSendRtpPacket(RTC::Consumer* consumer, RTC::RtpPacket* packet)
 	{
 		MS_TRACE();
 
@@ -2469,25 +2472,25 @@ namespace RTC
 				}
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(consumer, packet, cb);
 #else
 			const auto* cb = new onSendCallback([tccClient, &packetInfo](bool sent) {
 				if (sent)
 					tccClient->PacketSent(packetInfo, DepLibUV::GetTimeMsInt64());
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(consumer, packet, cb);
 #endif
 		}
 		else
 		{
-			SendRtpPacket(packet);
+			SendRtpPacket(consumer, packet);
 		}
 
 		this->sendRtpTransmission.Update(packet);
 	}
 
-	inline void Transport::OnConsumerRetransmitRtpPacket(RTC::Consumer* /*consumer*/, RTC::RtpPacket* packet)
+	inline void Transport::OnConsumerRetransmitRtpPacket(RTC::Consumer* consumer, RTC::RtpPacket* packet)
 	{
 		MS_TRACE();
 
@@ -2537,19 +2540,19 @@ namespace RTC
 				}
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(consumer, packet, cb);
 #else
 			const auto* cb = new onSendCallback([tccClient, &packetInfo](bool sent) {
 				if (sent)
 					tccClient->PacketSent(packetInfo, DepLibUV::GetTimeMsInt64());
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(consumer, packet, cb);
 #endif
 		}
 		else
 		{
-			SendRtpPacket(packet);
+			SendRtpPacket(consumer, packet);
 		}
 
 		this->sendRtxTransmission.Update(packet);
@@ -2638,30 +2641,7 @@ namespace RTC
 	{
 		MS_TRACE();
 
-		switch (dataConsumer->GetType())
-		{
-			case RTC::DataConsumer::Type::SCTP:
-			{
-				this->sctpAssociation->SendSctpMessage(dataConsumer, ppid, msg, len);
-
-				break;
-			}
-
-			case RTC::DataConsumer::Type::DIRECT:
-			{
-				// Notify the Node DirectTransport.
-				json data = json::object();
-
-				data["ppid"] = ppid;
-
-				PayloadChannel::Notifier::Emit(dataConsumer->id, "message", data, msg, len);
-
-				// Increase send transmission.
-				DataSent(len);
-
-				break;
-			}
-		}
+		SendMessage(dataConsumer, ppid, msg, len);
 	}
 
 	inline void Transport::OnDataConsumerDataProducerClosed(RTC::DataConsumer* dataConsumer)
@@ -2874,14 +2854,14 @@ namespace RTC
 				}
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(nullptr, packet, cb);
 #else
 			const auto* cb = new onSendCallback([tccClient, &packetInfo](bool sent) {
 				if (sent)
 					tccClient->PacketSent(packetInfo, DepLibUV::GetTimeMsInt64());
 			});
 
-			SendRtpPacket(packet, cb);
+			SendRtpPacket(nullptr, packet, cb);
 #endif
 		}
 		else
@@ -2889,7 +2869,7 @@ namespace RTC
 			// May emit 'trace' event.
 			EmitTraceEventProbationType(packet);
 
-			SendRtpPacket(packet);
+			SendRtpPacket(nullptr, packet);
 		}
 
 		this->sendProbationTransmission.Update(packet);